Set prelude docs to be no_inline, preventing dead links
The default seems to be that structs get docs repeated, whereas enums are shown as "re-exports". However, this means that the link from prelude/struct.LocalExecutor.html to prelude/enum.Latency.html is dead, because the latter does not exist.

This change means that all of them show up as re-exports.

We could also do the complete opposite with #[doc(inline)], repeating docs for both structs and enums in prelude. That would also prevent dead links.